+++ app/pb_kits/playbook/pb_stat_change/docs/_stat_change_default.jsx
@@ -20,7 +20,7 @@ const StatChangeDefault = () => {
 
       <StatChange
           change="neutral"
-          value={98}
+          value={102}
       />
     </div>
   )